Add user profile image addition
