+++ app/src/main/res/layout/activity_main.xml
@@ -33,12 +33,7 @@
                 android:text="register"
         />
 
-        <Button
-                android:id="@+id/img"
-                android:layout_height="wrap_content"
-                android:layout_width="wrap_content"
-                android:text="img"
-        />
+
