+++ app/src/main/res/layout/fragment_notifications.xml
@@ -17,4 +17,21 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"/>
+
+
+    <Button
+            android:id="@+id/img"
+            android:layout_height="wrap_content"
+            android:layout_width="wrap_content"
+            android:text="img"
+            tools:ignore="MissingConstraints"/>
+
+
+    <include
+            android:id="@+id/image_layout"
+            android:visibility="gone"
+            layout="@layout/image"
+    />
+
+
 </androidx.constraintlayout.widget.ConstraintLayout>
